Refactor and check for potential measurability of const
